wip: added EmployeeController that supports HTTP GET request and did some refactoring to other clases

+++ src/main/java/com/hpe/restful/resource/Employee.java
@@ -1,17 +1,17 @@
 package com.hpe.restful.resource;
 
 public class Employee {
-    private Long employee_id;
+    private int employee_id;
     private String first_name;
     private String last_name;
     private String email;
     private String title;
 
-    public Long getEmployee_id() {
+    public int getEmployee_id() {
         return employee_id;
     }
 
-    public void setEmployee_id(Long employee_id) {
+    public void setEmployee_id(int employee_id) {
         this.employee_id = employee_id;
     }
